requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-storage-queue
azure-data-tables
azure-identity
azure-keyvault-secrets
azure-servicebus>=7.11.0
geopandas
psycopg[binary]
rasterio>=1.3.1
rio-cogeo
rio-stac>=0.10.0

# STRONG TYPING DISCIPLINE - Schema validation and enforcement
pydantic>=2.0.0

# STAC Implementation - PgSTAC database and Python libraries
pypgstac==0.8.5
psycopg-pool>=3.1.0
pystac>=1.13.0
pystac-client>=0.7.0
stac-pydantic>=3.4.0